requirements.txt
langchain
langchain-community
langchain-chroma
chromadb
bs4
requests
python-dotenv
openai